// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  AGENTE: 'AGENTE',
  EMPLEADO: 'EMPLEADO'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const TicketStatus = {
  ABIERTO: 'ABIERTO',
  EN_PROCESO: 'EN_PROCESO',
  RESUELTO: 'RESUELTO',
  CERRADO: 'CERRADO'
} as const

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus]


export const Priority = {
  BAJA: 'BAJA',
  MEDIA: 'MEDIA',
  ALTA: 'ALTA'
} as const

export type Priority = (typeof Priority)[keyof typeof Priority]


export const NotificationType = {
  TICKET_ASIGNADO: 'TICKET_ASIGNADO',
  TICKET_RESUELTO: 'TICKET_RESUELTO',
  TICKET_CERRADO: 'TICKET_CERRADO',
  NUEVO_COMENTARIO: 'NUEVO_COMENTARIO'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]